+++ src/OrderProvider.tsx
@@ -1,24 +1,34 @@
-import React, { Fragment, ReactNode, useContext, useEffect, useState } from 'react';
-import AuthContext from "./AuthContext";
-import OrdersContext from "./OrdersContext";
-import ConfigContext, { IConfig } from "./ConfigContext";
-import * as orderingCore from "@orderingstack/ordering-core";
-import { INotificationMessage, ISteeringCommand } from '@orderingstack/ordering-types';
+import React, {
+  Fragment,
+  ReactNode,
+  useContext,
+  useEffect,
+  useState,
+} from 'react';
+import AuthContext from './AuthContext';
+import OrdersContext from './OrdersContext';
+import ConfigContext, { IConfig } from './ConfigContext';
+import * as orderingCore from '@orderingstack/ordering-core';
+import {
+  INotificationMessage,
+  ISteeringCommand,
+} from '@orderingstack/ordering-types';
 
 export interface OrderProviderProps {
   children: ReactNode;
-  onWebsocketNotification?: (message:INotificationMessage) => void;
-  onSteeringCommand?: (command: ISteeringCommand) => void;}
+  onWebsocketNotification?: (message: INotificationMessage) => void;
+  onSteeringCommand?: (command: ISteeringCommand) => void;
+}
 
 export default function OrderProvider(props: OrderProviderProps) {
   const _configContext = useContext(ConfigContext);
   if (!_configContext) {
-    return (<Fragment>Configuration should be provided</Fragment>)
+    return <Fragment>Configuration should be provided</Fragment>;
   }
   const config: IConfig = _configContext;
 
   const [orders, setOrders] = useState<any>({});
-  const { authProvider, loggedIn } = useContext( AuthContext);
+  const { authProvider, loggedIn } = useContext(AuthContext);
 
   const onOrdersUpdated = (order: any, allOrders: any) => {
     //console.log("--------------" + Object.keys(allOrders).length);
@@ -26,12 +36,15 @@ export default function OrderProvider(props: OrderProviderProps) {
   };
 
   async function setOrderChangesListener(
-    authProvider: orderingCore.IConfiguredAuthDataProvider
+    authProvider: orderingCore.IConfiguredAuthDataProvider,
   ) {
+    let disconnect: () => Promise<void> = () => {
+      return Promise.resolve();
+    };
     try {
       const { token } = await authProvider();
       if (token) {
-        orderingCore.orderChangesListener(
+        disconnect = await orderingCore.orderChangesListener(
           config.baseUrl,
           config.tenant,
           config.venue,
@@ -42,17 +55,21 @@ export default function OrderProvider(props: OrderProviderProps) {
           },
           config.enableKDS,
           props.onWebsocketNotification,
-          props.onSteeringCommand
+          props.onSteeringCommand,
         );
       }
     } catch (err) {
       setOrders({});
     }
+    return disconnect;
   }
 
   useEffect(() => {
     orderingCore.setOrderStoreUpdatedCallback(onOrdersUpdated);
-    setOrderChangesListener(authProvider);
+    const promise = setOrderChangesListener(authProvider);
+    return () => {
+      promise.then((disconnect) => disconnect());
+    };
   }, [loggedIn]);
 
   return (